Homework 10.2 Удалил файл widget.py вместо него сделал файл widg.py перенёс код туда, наконец то заработали тесты и проблема с импортами ушла....
