+++ src/processing.py
@@ -1,4 +1,4 @@
-list = [
+my_list = [
     {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
     {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
     {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
@@ -21,4 +21,4 @@ def sort_by_date(my_list: list, reverse=True) -> list:
     return sort_date_list
 
 
-print(filter_by_state(list))
+print(filter_by_state(my_list))
