+++ tests/test_mask.py
@@ -1,8 +1,8 @@
 import pytest
-from src import *
-from src.masks import *
+
+from src.masks import get_mask_account, get_mask_card_number
 from src.processing import filter_by_state
-from src.widget import *
+from src.widg import funk, get_date
 
 
 @pytest.mark.parametrize(
@@ -15,6 +15,7 @@ from src.widget import *
     ],
 )
 def test_get_mask_card_number(number_card, mask):
+    """Тест функции get_mask_card_number"""
     assert get_mask_card_number(number_card) == mask
 
 
@@ -23,23 +24,26 @@ def test_get_mask_card_number(number_card, mask):
     [("12345678901234567890", "** 7890"), ("98765432109876543210", "** 3210"), ("", "Некорректный номер")],
 )
 def test_get_mask_account(number_account, mask):
+    """Тест функции get_mask_account"""
     assert get_mask_account(number_account) == mask
 
 
-# @pytest.mark.parametrize(
-#     "inf_the_card, mask",
-#     [
-#         ("Visa Platinum 1234567890123456", " Visa Platinum 1234 56** **** 3456"),
-#         ("Maestro 9876543210123456", " Maestro 9876 54** **** 3456"),
-#         ("Счет 1234567890123456790", "Некорректный номер счёта"),
-#         ("Счет 12345678901234567890", " Счет ** 7890"),
-#     ],
-# )
-# def test_mask_account_card(inf_the_card: str, mask: str) -> str:
-#     assert f(inf_the_card) == mask
+@pytest.mark.parametrize(
+    "inf_the_card, mask",
+    [
+        ("Visa Platinum 1234567890123456", " Visa Platinum 1234 56** **** 3456"),
+        ("Maestro 9876543210123456", " Maestro 9876 54** **** 3456"),
+        ("Счет 1234567890123456790", "Некорректный номер счёта"),
+        ("Счет 12345678901234567890", "Счет ** 7890"),
+    ],
+)
+def test_mask_account_card(inf_the_card: str, mask: str) -> str:
+    """Тест функции funk"""
+    assert funk(inf_the_card) == mask
 
 
 def test_get_mask_card(get_mask):
+    """Тест функции get_mask_card_number"""
     assert get_mask_card_number("1234567890123456") == get_mask
 
 
@@ -59,10 +63,12 @@ def test_get_mask_card(get_mask):
     ],
 )
 def test_filter_by_state(list_dictionaries, state):
+    """Тест функции filter_by_state"""
     assert filter_by_state(list_dictionaries) == state
 
 
 def test_get_date(get_data):
+    """Тест функции get_date"""
     assert get_date("2024-03-11T02:26:18.671407") == get_data
 
 
@@ -75,5 +81,6 @@ def test_get_date(get_data):
         ("2023-04-11T02:26:18.6123", "11.04.2023"),
     ],
 )
-def test_get_date(date, targ):
+def test_get_date_param(date, targ):
+    """Тест функции get_date"""
     assert get_date(date) == targ
